Refactored express router
